mono/AI_test/requirements.txt
flask
google-generativeai
pypdf2
yt-dlp